[map-helper-functions] Start of helper functions with focus on generateRuntimeLegend.ts
