+++ packages/map/src/helpers/setBinNumbers.ts
@@ -1,0 +1,5 @@
+export const setBinNumbers = result => {
+  result.items.forEach((row, i) => {
+    row.bin = i // set bin number to index
+  })
+}
